Clean rewrite of insilicoSV
Rewrote from scratch with a clean design. SVs are modeled as lists of numbered breakends and operations in terms of these breakends. In SV construction phase, SVs (unplaced) are instantiated from configs. In a separate placement phase, breakends are assigned to genomic locations, currently for one SV at a time (but with room to implement globally optimized placement later).  Operations are classified as either in-place (transforming a piece of reference in-place) or new-place (inserting a possibly transformed piece of reference or novel sequence at a new location), and all transforms are represented in terms of three basic transforms (identity, inversion or deletion).