[FEATURE] Enable overlapping heterozygous SVs across distinct haplotypes. (#20)
Allows for cases where an SV exists on one haplotype and another SV exists on the homologous haplotype, with their genomic regions overlapping.
To efficiently keep track of available genomic regions, an interval tree is now used for each haplotype, and an additional tree maintains the intersection of these regions for homozygous variants.